requirements.txt
numpy==1.25.2
six==1.16.0
azure-core==1.29.4
azure-iot-device==2.12.0
azure-iot-hub==2.6.1